hide recruiting packages
